# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Law Firm Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: clients
    description: Client operations
  - name: cases
    description: Case operations
  - name: invoices
    description: Invoice operations
  - name: tasks
    description: Task operations
  - name: events
    description: Calendar event operations
  - name: documents
    description: Document operations
  - name: dashboard
    description: Dashboard summary operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ---- DASHBOARD ----
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/today:
    get:
      operationId: getDashboardToday
      tags: [dashboard]
      summary: Get today's events and tasks
      responses:
        "200":
          description: Today overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayOverview"

  /dashboard/alerts:
    get:
      operationId: getDashboardAlerts
      tags: [dashboard]
      summary: Get upcoming deadline alerts
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

  # ---- CLIENTS ----
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- CASES ----
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all cases
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: court
          in: query
          schema:
            type: string
        - name: lawyerId
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "201":
          description: Created case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"

  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    put:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- INVOICES ----
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- TASKS ----
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: caseId
          in: query
          schema:
            type: integer
        - name: done
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- EVENTS ----
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List calendar events
      parameters:
        - name: caseId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"

  /events/{id}:
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- DOCUMENTS ----
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: caseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    DashboardSummary:
      type: object
      properties:
        activeCases:
          type: integer
        monthlyIncome:
          type: number
        pendingInvoices:
          type: integer
        upcomingDeadlines:
          type: integer
      required: [activeCases, monthlyIncome, pendingInvoices, upcomingDeadlines]

    TodayOverview:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/CalendarEvent"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
      required: [sessions, tasks]

    Alert:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
        type:
          type: string
          enum: [deadline, hearing, payment]
        dueDate:
          type: string
          format: date
        caseId:
          type: integer
          nullable: true
        caseName:
          type: string
          nullable: true
      required: [id, message, type, dueDate]

    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]

    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name]

    Case:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        status:
          type: string
          enum: [active, closed, suspended, pending]
        court:
          type: string
          nullable: true
        lawyer:
          type: string
          nullable: true
        nextHearing:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, clientId, clientName, status, createdAt]

    CreateCaseBody:
      type: object
      properties:
        title:
          type: string
        clientId:
          type: integer
        status:
          type: string
          enum: [active, closed, suspended, pending]
        court:
          type: string
          nullable: true
        lawyer:
          type: string
          nullable: true
        nextHearing:
          type: string
          format: date
          nullable: true
        description:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [title, clientId, status]

    Invoice:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        caseId:
          type: integer
          nullable: true
        caseName:
          type: string
          nullable: true
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled]
        dueDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, clientId, clientName, amount, status, createdAt]

    CreateInvoiceBody:
      type: object
      properties:
        clientId:
          type: integer
        caseId:
          type: integer
          nullable: true
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, cancelled]
        dueDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
      required: [clientId, amount, status]

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        caseId:
          type: integer
          nullable: true
        caseName:
          type: string
          nullable: true
        done:
          type: boolean
        dueDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, done, createdAt]

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        caseId:
          type: integer
          nullable: true
        done:
          type: boolean
        dueDate:
          type: string
          format: date
          nullable: true
      required: [title]

    CalendarEvent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        caseId:
          type: integer
          nullable: true
        caseName:
          type: string
          nullable: true
        date:
          type: string
          format: date
        time:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        type:
          type: string
          enum: [hearing, meeting, deadline, other]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, date, type, createdAt]

    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        caseId:
          type: integer
          nullable: true
        date:
          type: string
          format: date
        time:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        type:
          type: string
          enum: [hearing, meeting, deadline, other]
        notes:
          type: string
          nullable: true
      required: [title, date, type]

    Document:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        caseId:
          type: integer
          nullable: true
        caseName:
          type: string
          nullable: true
        fileType:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]

    CreateDocumentBody:
      type: object
      properties:
        name:
          type: string
        caseId:
          type: integer
          nullable: true
        fileType:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
      required: [name]